requirements.txt
opencv-python
fer
speechrecognition
pyttsx3
groq
numpy
pyaudio
python-dotenv
tensorflow
setuptools
textblob
flask
flask-socketio
langdetect
gTTS
